Wrap the CAN include/linux header files to include/socketcan
To avoid conflicts with the *almost* similar kernel include files we
use a separate set of header files in include/socketcan (instead of
include/linux). But the include/linux header files are still needed
for the user space applications, e.g. the can-utils. To simplify
maintanance, wrappers to the corresponding files in include/socketcan
are used. Kernel-only header files are removed from include/linux.

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>


git-svn-id: svn://svn.berlios.de/socketcan/trunk@1070 030b6a49-0b11-0410-94ab-b0dab22257f2