can: fix WARN_ON dump in net/core/rtnetlink.c:rtmsg_ifinfo()
On older kernels, e.g. 2.6.27, a WARN_ON dump in rtmsg_ifinfo()
is thrown when the CAN device is registered due to insufficient
skb space, as reported by various users. This patch adds the
rtnl_link_ops "get_size" to fix the problem. I think it's
required for more recent kernels as well, even if no WARN_ON
dumps are triggered. Maybe we also need it for the CAN xstats.

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>



git-svn-id: svn://svn.berlios.de/socketcan/trunk@1078 030b6a49-0b11-0410-94ab-b0dab22257f2